Move ShaderTranslation.cpp/h to Common/GPU.
