+++ Common/Common.vcxproj
@@ -430,6 +430,7 @@
     <ClInclude Include="GPU\OpenGL\GLCommon.h" />
     <ClInclude Include="GPU\OpenGL\GLDebugLog.h" />
     <ClInclude Include="GPU\Shader.h" />
+    <ClInclude Include="GPU\ShaderTranslation.h" />
     <ClInclude Include="GPU\ShaderWriter.h" />
     <ClInclude Include="GPU\thin3d.h" />
     <ClInclude Include="GPU\thin3d_create.h" />
@@ -753,6 +754,7 @@
     <ClCompile Include="GPU\OpenGL\GLDebugLog.cpp" />
     <ClCompile Include="GPU\OpenGL\thin3d_gl.cpp" />
     <ClCompile Include="GPU\Shader.cpp" />
+    <ClCompile Include="GPU\ShaderTranslation.cpp" />
     <ClCompile Include="GPU\ShaderWriter.cpp" />
     <ClCompile Include="GPU\thin3d.cpp" />
     <ClCompile Include="GPU\Vulkan\thin3d_vulkan.cpp" />
