+++ Common/GPU/ShaderTranslation.cpp
@@ -38,7 +38,7 @@
 #include "Common/StringUtils.h"
 #include "Common/GPU/Shader.h"
 
-#include "GPU/Common/ShaderTranslation.h"
+#include "Common/GPU/ShaderTranslation.h"
 #include "ext/glslang/SPIRV/GlslangToSpv.h"
 #include "Common/GPU/thin3d.h"
 #include "Common/GPU/OpenGL/GLFeatures.h"
