+++ GPU/GPU.vcxproj.filters
@@ -222,9 +222,6 @@
     <ClInclude Include="D3D11\D3D11Util.h">
       <Filter>D3D11</Filter>
     </ClInclude>
-    <ClInclude Include="Common\ShaderTranslation.h">
-      <Filter>Common</Filter>
-    </ClInclude>
     <ClInclude Include="Software\Sampler.h">
       <Filter>Software</Filter>
     </ClInclude>
@@ -484,9 +481,6 @@
     </ClCompile>
     <ClCompile Include="D3D11\StencilBufferD3D11.cpp">
       <Filter>D3D11</Filter>
-    </ClCompile>
-    <ClCompile Include="Common\ShaderTranslation.cpp">
-      <Filter>Common</Filter>
     </ClCompile>
     <ClCompile Include="Common\ShaderCommon.cpp">
       <Filter>Common</Filter>
