+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -420,6 +420,7 @@
     <ClInclude Include="..\..\Common\GPU\DataFormat.h" />
     <ClInclude Include="..\..\Common\GPU\OpenGL\GLFeatures.h" />
     <ClInclude Include="..\..\Common\GPU\Shader.h" />
+    <ClInclude Include="..\..\Common\GPU\ShaderTranslation.h" />
     <ClInclude Include="..\..\Common\GPU\ShaderWriter.h" />
     <ClInclude Include="..\..\Common\GPU\thin3d.h" />
     <ClInclude Include="..\..\Common\GPU\thin3d_create.h" />
@@ -539,6 +540,7 @@
     <ClCompile Include="..\..\Common\GPU\D3D11\thin3d_d3d11.cpp" />
     <ClCompile Include="..\..\Common\GPU\OpenGL\GLFeatures.cpp" />
     <ClCompile Include="..\..\Common\GPU\Shader.cpp" />
+    <ClCompile Include="..\..\Common\GPU\ShaderTranslation.cpp" />
     <ClCompile Include="..\..\Common\GPU\ShaderWriter.cpp" />
     <ClCompile Include="..\..\Common\GPU\thin3d.cpp" />
     <ClCompile Include="..\..\Common\Input\GestureDetector.cpp" />
