+++ libretro/Makefile.common
@@ -168,6 +168,7 @@ SOURCES_CXX += \
 	$(COMMONDIR)/GPU/thin3d.cpp \
 	$(COMMONDIR)/GPU/Shader.cpp \
 	$(COMMONDIR)/GPU/ShaderWriter.cpp \
+	$(COMMONDIR)/GPU/ShaderTranslation.cpp \
 	$(COMMONDIR)/GPU/OpenGL/thin3d_gl.cpp \
 	$(COMMONDIR)/GPU/OpenGL/GLDebugLog.cpp \
 	$(COMMONDIR)/GPU/OpenGL/GLSLProgram.cpp \
@@ -238,7 +239,6 @@ SOURCES_CXX += \
 	$(GPUCOMMONDIR)/ShaderId.cpp \
 	$(GPUCOMMONDIR)/ShaderCommon.cpp \
 	$(GPUCOMMONDIR)/ShaderUniforms.cpp \
-	$(GPUCOMMONDIR)/ShaderTranslation.cpp \
 	$(GPUCOMMONDIR)/GPUDebugInterface.cpp \
 	$(GPUCOMMONDIR)/DepalettizeShaderCommon.cpp \
 	$(GPUCOMMONDIR)/TransformCommon.cpp \
